refactor: Complete Phase 5B.1 - Split cli.py into modules
- Create src/traceflux/cli/ package structure
- Extract commands to separate modules:
  - search.py: Search command implementation
  - index.py: Index command implementation
  - patterns.py: Patterns command implementation
  - associations.py: Associations command implementation
- Extract helpers to shared module:
  - scan_paths: Path scanning utility
  - show_suggestions: Association suggestions
- Create main.py for argument parsing
- Add __init__.py and __main__.py for package execution
- Reduce main cli.py from 602 lines to ~180 lines
- Each command module <150 lines (focused scope)
- Full backward compatibility maintained
- All functionality preserved

Closes: #25 (Phase 5B.1: Refactor cli.py - Split into modules)
