+++ src/traceflux/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI package - Command-line interface for traceflux."""
+
+from .main import main
+
+__all__ = ["main"]
