+++ src/traceflux/cli/__main__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+"""Allow running traceflux.cli as a module: python -m traceflux.cli"""
+
+import sys
+from traceflux.cli.main import main
+
+if __name__ == "__main__":
+    sys.exit(main())
